requirements.txt
absl-py
cvxpy~=1.5.2
cvxopt==1.3.2
dash~=2.18.2
filterpy~=1.4.5
gitpython
ml_collections
numpy==1.23.1
pybullet~=3.2.6
qpth~=0.0.18
omegaconf==2.3.0
scikit-fmm~=2024.5.29
scikit-image~=0.21.0
tensorboard
torch~=2.1.2
torchaudio
torchvision
tensorflow==2.13.0
tensorflow-probability==0.21.0
tqdm~=4.67.1
matplotlib~=3.7.5
opencv-python~=4.10.0.84
open3d~=0.19.0
scipy~=1.10.1
plotly~=5.24.1
pyyaml~=6.0.2
pandas~=2.0.3
setuptools~=75.3.0
pillow~=10.4.0
imageio~=2.35.1